Removed RemoteContactsProvider, moved code to PubLookup; Using PubLookup;
